+++ filter/vibrato.go
@@ -2,5 +2,5 @@ package filter
 
 type Vibrato struct {
 	Frequency float32 `json:"frequency,omitempty"`
-	Depth     uint    `json:"depth,omitempty"`
+	Depth     float32 `json:"depth,omitempty"`
 }
